// src/api/models/RegistrarEstadoComponenteRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RegistrarEstadoComponenteRequest = {
    parteSubtipoId: number;
    estadoComponente: string;
    'severidadDaño': string;
    descripcion?: string | null;
};